from lecture

+++ src/main/java/ru/vk/itmo/osipovdaniil/DiskStorageUtils.java
@@ -4,6 +4,7 @@ import ru.vk.itmo.BaseEntry;
 import ru.vk.itmo.Entry;
 
 import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.lang.foreign.Arena;
 import java.lang.foreign.MemorySegment;
 import java.lang.foreign.ValueLayout;
@@ -14,13 +15,12 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardOpenOption;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.*;
+import java.util.stream.Stream;
 
 public final class DiskStorageUtils {
 
+    public static final String SSTABLE_PREFIX = "sstable_";
     private static final String INDEX = "index.idx";
     private static final String INDEX_TMP = "index.tmp";
 
@@ -56,15 +56,16 @@ public final class DiskStorageUtils {
     }
 
     public static List<MemorySegment> loadOrRecover(final Path storagePath, final Arena arena) throws IOException {
+        if (Files.exists(compactionFile(storagePath))) {
+            finalizeCompaction(storagePath);
+        }
         final Path indexTmp = getIndexTmpPath(storagePath);
         final Path indexFile = getIndexPath(storagePath);
-        if (Files.exists(indexTmp)) {
-            Files.move(indexTmp, indexFile, StandardCopyOption.ATOMIC_MOVE, StandardCopyOption.REPLACE_EXISTING);
-        } else {
-            try {
+        if (!Files.exists(indexFile)) {
+            if (Files.exists(indexTmp)) {
+                Files.move(indexTmp, indexFile, StandardCopyOption.ATOMIC_MOVE, StandardCopyOption.REPLACE_EXISTING);
+            } else {
                 Files.createFile(indexFile);
-            } catch (FileAlreadyExistsException ignored) {
-                // it is ok, actually it is normal state
             }
         }
         final List<String> existedFiles = Files.readAllLines(indexFile, StandardCharsets.UTF_8);
@@ -202,7 +203,7 @@ public final class DiskStorageUtils {
             // it is ok, actually it is normal state
         }
         final List<String> existedFiles = Files.readAllLines(indexFile, StandardCharsets.UTF_8);
-        final String newFileName = String.valueOf(existedFiles.size());
+        final String newFileName = SSTABLE_PREFIX + existedFiles.size();
         long dataSize = 0;
         long count = 0;
         for (Entry<MemorySegment> entry : iterable) {
@@ -260,15 +261,125 @@ public final class DiskStorageUtils {
                 }
             }
         }
-        Files.move(indexFile, indexTmp, StandardCopyOption.ATOMIC_MOVE, StandardCopyOption.REPLACE_EXISTING);
         final List<String> list = new ArrayList<>(existedFiles.size() + 1);
         list.addAll(existedFiles);
         list.add(newFileName);
-        Files.write(indexFile,
+        Files.write(indexTmp,
                 list,
                 StandardOpenOption.WRITE,
                 StandardOpenOption.CREATE,
                 StandardOpenOption.TRUNCATE_EXISTING);
-        Files.delete(indexTmp);
+        Files.deleteIfExists(indexFile);
+        Files.move(indexFile, indexTmp, StandardCopyOption.ATOMIC_MOVE);
+    }
+
+    public static void compact(final Path storagePath, final Iterable<Entry<MemorySegment>> iterable)
+            throws IOException {
+        final String newFileName = "compaction.tmp";
+        final Path compactionTmpFile = storagePath.resolve(newFileName);
+        long dataSize = 0;
+        long count = 0;
+        for (Entry<MemorySegment> entry : iterable) {
+            dataSize += entry.key().byteSize();
+            MemorySegment value = entry.value();
+            if (value != null) {
+                dataSize += value.byteSize();
+            }
+            count++;
+        }
+        long indexSize = count * 2 * Long.BYTES;
+        try (FileChannel fileChannel = FileChannel.open(
+                compactionTmpFile,
+                StandardOpenOption.WRITE,
+                StandardOpenOption.READ,
+                StandardOpenOption.CREATE);
+                Arena writeArena = Arena.ofConfined()) {
+            final MemorySegment fileSegment = fileChannel.map(
+                    FileChannel.MapMode.READ_WRITE,
+                    0,
+                    indexSize + dataSize,
+                    writeArena
+            );
+
+            // index:
+            // |key0_Start|value0_Start|key1_Start|value1_Start|key2_Start|value2_Start|...
+            // key0_Start = data start = end of index
+            long dataOffset = indexSize;
+            int indexOffset = 0;
+            for (Entry<MemorySegment> entry : iterable) {
+                fileSegment.set(ValueLayout.JAVA_LONG_UNALIGNED, indexOffset, dataOffset);
+                dataOffset += entry.key().byteSize();
+                indexOffset += Long.BYTES;
+
+                MemorySegment value = entry.value();
+                if (value == null) {
+                    fileSegment.set(ValueLayout.JAVA_LONG_UNALIGNED, indexOffset, tombstone(dataOffset));
+                } else {
+                    fileSegment.set(ValueLayout.JAVA_LONG_UNALIGNED, indexOffset, dataOffset);
+                    dataOffset += value.byteSize();
+                }
+                indexOffset += Long.BYTES;
+            }
+
+            // data:
+            // |key0|value0|key1|value1|...
+            dataOffset = indexSize;
+            for (Entry<MemorySegment> entry : iterable) {
+                MemorySegment key = entry.key();
+                MemorySegment.copy(key, 0, fileSegment, dataOffset, key.byteSize());
+                dataOffset += key.byteSize();
+
+                MemorySegment value = entry.value();
+                if (value != null) {
+                    MemorySegment.copy(value, 0, fileSegment, dataOffset, value.byteSize());
+                    dataOffset += value.byteSize();
+                }
+            }
+        }
+        Files.move(compactionTmpFile,
+                storagePath.resolve("compaction"),
+                StandardCopyOption.ATOMIC_MOVE,
+                StandardCopyOption.REPLACE_EXISTING);
+        finalizeCompaction(storagePath);
+    }
+
+    private static void finalizeCompaction(Path storagePath) throws IOException {
+        Path compactionFile = compactionFile(storagePath);
+        try (Stream<Path> stream = Files.find(storagePath, 1, (path, _) -> path.getFileName().toString().startsWith(SSTABLE_PREFIX))) {
+            stream.forEach(p -> {
+                try {
+                    Files.delete(p);
+                } catch (IOException e) {
+                    throw new UncheckedIOException(e);
+                }
+            });
+        }
+
+        Path indexTmp = storagePath.resolve("index.tmp");
+        Path indexFile = storagePath.resolve("index.idx");
+
+        Files.deleteIfExists(indexFile);
+        Files.deleteIfExists(indexTmp);
+
+        boolean noData = Files.size(compactionFile) == 0;
+
+        Files.write(
+                indexTmp,
+                noData ? Collections.emptyList() : Collections.singleton(SSTABLE_PREFIX + "0"),
+                StandardOpenOption.WRITE,
+                StandardOpenOption.CREATE,
+                StandardOpenOption.TRUNCATE_EXISTING
+        );
+
+        Files.move(indexTmp, indexFile, StandardCopyOption.ATOMIC_MOVE);
+        if (noData) {
+            Files.delete(compactionFile);
+        } else {
+            Files.move(compactionFile, storagePath.resolve(SSTABLE_PREFIX + "0"), StandardCopyOption.ATOMIC_MOVE);
+        }
+    }
+
+    private static Path compactionFile(Path storagePath) {
+        return storagePath.resolve("compaction");
     }
 }
